1：新增Backup、Copy类； 2：修改分页long转换问题； 3：setWhere表达式bug修复； 4：创建表时忽略掉需要忽略的字段； 5：修改updatetime注解应在每次更新时插入最新时间； 6：修改逻辑删除不应带上乐观锁条件； 7：增加AbstractSchemaAspect类统一设置schema； 8：修改back备份数据默认表名时间格式化到毫秒； 9：copy，back等增加schema支持；

+++ vonce-sqlbean-core/src/main/java/cn/vonce/sql/page/PageHelper.java
@@ -251,10 +251,10 @@ public class PageHelper<T> {
             int count;
             Object obj;
             if (tClazz != null) {
-                count = ReflectUtil.instance().invoke(clazz, pageService, this.getPagingMethod().getCount(), new Class[]{Class.class, Select.class}, new Object[]{tClazz, select});
+                count = (int) ReflectUtil.instance().invoke(clazz, pageService, this.getPagingMethod().getCount(), new Class[]{Class.class, Select.class}, new Object[]{tClazz, select});
                 obj = ReflectUtil.instance().invoke(clazz, pageService, this.getPagingMethod().getSelect(), new Class[]{Class.class, Select.class}, new Object[]{tClazz, sqlBeanSelect});
             } else {
-                count = ReflectUtil.instance().invoke(clazz, pageService, this.getPagingMethod().getCount(), new Class[]{Select.class}, new Object[]{select});
+                count = (int) ReflectUtil.instance().invoke(clazz, pageService, this.getPagingMethod().getCount(), new Class[]{Select.class}, new Object[]{select});
                 obj = ReflectUtil.instance().invoke(clazz, pageService, this.getPagingMethod().getSelect(), new Class[]{Select.class}, new Object[]{sqlBeanSelect});
             }
             // 计算共有几页
